Markdown reader: detect unused footnotes
Here's a toy implementation dealing with #1718 for Markdown. Adds a
field to ParserState to track footnote references.

This issue affects all readers. Since the Pandoc intermediate format
inlines all footnotes, we can't perform a single sanity check after
reading. So, we would need to implement similar code for each.
That might not be desirable.